pyproject.toml
[tool.poetry]
name = "stateapi"
version = "0.1.13"
description = ""
authors = ["wwqdrh <huiloademail@163.com>"]
readme = "README.md"
packages = [{include = "wwqdrh/stateapi"}]

[tool.poetry.dependencies]
python = "^3.10"
pykit = {git = "https://github.com/wwqdrh/pykit.git", rev = "v0.0.2"}
peewee = "^3.15.3"
fastapi = "^0.85.2"
uvicorn = "^0.19.0"
grpcio-tools = "^1.51.1"
protobuf = "^4.21.12"
grpclib = {extras = ["protobuf"], version = "^0.4.3"}


[tool.poetry.group.dev.dependencies]
black = "^22.10.0"
pytest = "^7.2.0"
pytest-asyncio = "^0.20.1"
mypy = "^0.991"
grpcio-tools = "^1.51.1"
protobuf = "^4.21.12"
grpclib = {extras = ["protobuf"], version = "^0.4.3"}


[tool.poetry.group.build.dependencies]
nuitka = "^1.3.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
